Fixed transform bug by adding tests Adjusted camera projection Added duplcation and suppression of game objects Changed rotation type to double instead of float

+++ src/main/kotlin/physics2d/components/Box2DCollider.kt
@@ -10,6 +10,6 @@ class Box2DCollider: Collider() {
 
     override fun editorUpdate(dt: Float) {
         val center = Vector2f(gameObject.transform.position).add(offset)
-        DebugDraw.addBox2D(center, halfSize, gameObject.transform.rotation.toDouble())
+        DebugDraw.addBox2D(center, halfSize, gameObject.transform.rotation)
     }
 }
